FetchURLTask to make a get request asynchronously
FetchURLTask performs asynchronous task of fetching data from given URL. It
accepts two arguments in its constructor: FetchURLTaskPostExecution and Converter.
FetchURLTaskPostExecution's is a interface with one 'execute' method. This method is
automatically called when the task finishes with the data fetched from URL which
is 'converted' to the given type using the Converter interface. Converter interface
is also a interface with one 'convert' method. 'convert' method is called at the
end of 'doInBackground' method of FetchURLTask. If Converter argument happens to be null,
then the default converter is used which is no converter at all. It just returns plain
String object.